Add documentation, bring library up to date with working installed version and tidy up
Moved the LICENSE file to Resources/meta/LICENSE which seems to be best
practice
Remove the EnumTransactionType as it's not really part of this project.
Added some installation information to README.md. Should get a user to
a working state but I've not tested it from scratch using these
instructions yet - these are just notes of what I've had to do to fix
the implementation issues I've been having.

Modified the information being stored in config.yml for
stof_doctrine_extensions from the config in resources (where it caused
an almighty issue) to the calling application's config.yml. See the
documentation